NIFI-14778: Relocating ProcessGroupReferences component into top level common components so it can be referenced across pages.
