+++ nifi-frontend/src/main/frontend/apps/nifi/src/app/pages/parameter-contexts/feature/parameter-contexts.module.ts
@@ -26,7 +26,7 @@ import { ParameterContextListingEffects } from '../state/parameter-context-listi
 import { Navigation } from '../../../ui/common/navigation/navigation.component';
 import { BannerText } from '../../../ui/common/banner-text/banner-text.component';
 import { ParameterContextListing } from '../ui/parameter-context-listing/parameter-context-listing.component';
-import { ProcessGroupReferences } from '../ui/parameter-context-listing/process-group-references/process-group-references.component';
+import { ProcessGroupReferences } from '../../../ui/common/process-group-references/process-group-references.component';
 
 @NgModule({
     declarations: [ParameterContexts],
