+++ nifi-frontend/src/main/frontend/apps/nifi/src/app/ui/common/process-group-references/process-group-references.component.spec.ts
